+++ src/app/user-service.service.ts
@@ -1,12 +1,12 @@
 import { Injectable, inject, signal } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { BehaviorSubject, Observable, tap } from 'rxjs';
+import { BehaviorSubject, Observable, tap, of } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
 })
 export class UserServiceService {
-  private http = inject(HttpClient);
+  private http = inject(HttpClient, { optional: true });
   private apiUrl = 'http://localhost:3000/api';
 
   userList = signal<any[]>([]);
@@ -17,10 +17,13 @@ export class UserServiceService {
 
   fetchUsers() {
     const headers = this.getAuthHeaders();
-    this.http.get<any[]>(`${this.apiUrl}/users`, { headers }).subscribe(data => this.userList.set(data));
+    if (this.http) {
+      this.http.get<any[]>(`${this.apiUrl}/users`, { headers }).subscribe(data => this.userList.set(data));
+    }
   }
 
   saveUser(data: any): Observable<any> {
+    if (!this.http) return of(null) as Observable<any>;
     return this.http.post(`${this.apiUrl}/users`, data).pipe(
       tap((response: any) => {
         this.setSession(response.user, response.token);
@@ -30,6 +33,7 @@ export class UserServiceService {
   }
 
   login(email: string, password: string): Observable<any> {
+    if (!this.http) return of(null) as Observable<any>;
     return this.http.post(`${this.apiUrl}/login`, { email, password }).pipe(
       tap((response: any) => {
         this.setSession(response.user, response.token);
@@ -61,13 +65,17 @@ export class UserServiceService {
 
   deleteUser(id: string) {
     const headers = this.getAuthHeaders();
-    return this.http.delete(`${this.apiUrl}/users/${id}`, { headers }).subscribe(() => 
-      this.userList.update(list => list.filter(p => p._id !== id))
-    );
+    if (this.http) {
+      return this.http.delete(`${this.apiUrl}/users/${id}`, { headers }).subscribe(() =>
+        this.userList.update(list => list.filter(p => p._id !== id))
+      );
+    }
+    return null;
   }
 
   updateUser(id: string, data: any) {
     const headers = this.getAuthHeaders();
+    if (!this.http) return of(null) as Observable<any>;
     return this.http.put(`${this.apiUrl}/users/${id}`, data, { headers });
   }
